Reorganized src and test folders
