+++ src/blockchain/block.js
@@ -1,5 +1,5 @@
 const SHA256 = require('crypto-js/sha256');
-const { DIFFICULTY, MINE_RATE } = require('../config');
+const { DIFFICULTY, MINE_RATE } = require('../../config');
 
 class Block {
     constructor(timestamp, lastHash, hash, data, nonce, difficulty) {
